Choices List
Ok, so I'm stuck trying to get the inquirer choices list to populate as an array. I've been reading on it, and I feel like Im close to getting it. I could just have the user select by input, but that would be a shortcut and really having the choices list be created by an array would be nifty